+++ backend/aufguss_backend/aufguss_app/urls.py
@@ -1,0 +1,9 @@
+from rest_framework.routers import DefaultRouter
+from .views import UserViewSet, AufgussViewSet, ChatViewSet
+
+router = DefaultRouter()
+router.register(r'users', UserViewSet)
+router.register(r'aufguesse', AufgussViewSet)
+router.register(r'chats', ChatViewSet)
+
+urlpatterns = router.urls